Modify r0 and dptr initialization.
Modify r0 and dptr initialization, so that after the first dup r0=0, r1=0, dptr=0. This effectively utilizes as data stack the last two bytes of internal ram... just after we push the first value on the data stack ( dup,  value # dptr mov, ) r0 and r1 will be initialized with 0 and dptr will hold the value, after the next dup that value will be stored on 0xfe-0xff.